requirements.txt
essentia==2.1b6.dev1389
kafka-python==2.0.2
numpy==1.24.3
